Modularize negotiation tests and fix client doc links
