+++ docs/gaps.md
@@ -56,7 +56,7 @@ current repository state and can be regenerated by auditing the source tree.
   `[role=3.4.1-rust]` trailers, and normalise Rust `file!()` paths to
   repo-relative form for diagnostics.
 - **Client orchestration and local copy pipeline** (`crates/core::client` and
-  `crates/engine`): [`run_client`](../crates/core/src/client.rs) delegates to
+  `crates/engine`): [`run_client`](../crates/core/src/client/mod.rs) delegates to
   [`LocalCopyPlan`](../crates/engine/src/local_copy.rs) for deterministic local
   filesystem copies covering regular files, directories, and symbolic links,
   preserving permissions and timestamps while the full delta engine is under
